decrating: collapse dual implementation for git (preserve shim logic)
Eliminates the dual-path pattern for the git subsystem:
- Deletes the stale in-tree `git.rs` (the large monolithic copy).
- Promotes `git_micro.rs` to canonical `git.rs`.
- The shim's added logic is preserved:
  - `SHIPPER_GIT_BIN` environment override: when set, both
    `collect_git_context` and `is_git_clean` route through a local
    `Command`-based implementation that respects the custom git binary
    path. When unset, they delegate to `shipper_git` (the microcrate).
  - `git_program()` / `is_repo_root()` / `local_is_git_clean()`
    helpers backing the override path.
  - `ensure_git_clean` wrapper preserving the historical error
    message ("commit/stash changes or use --allow-dirty").
- Removes `micro-git` feature from shipper and shipper-cli Cargo.toml,
  makes `shipper-git` a non-optional dep, drops it from `micro-all`.
