+++ crates/shipper/Cargo.toml
@@ -16,7 +16,7 @@ categories = ["development-tools", "development-tools::build-utils"]
 # Workspace microcrates
 shipper-retry = { path = "../shipper-retry" }
 shipper-encrypt = { path = "../shipper-encrypt" }
-shipper-git = { path = "../shipper-git", optional = true }
+shipper-git = { path = "../shipper-git" }
 shipper-events = { path = "../shipper-events" }
 shipper-execution-core = { path = "../shipper-execution-core" }
 shipper-environment = { path = "../shipper-environment" }
@@ -61,7 +61,6 @@ hmac = "0.13"
 sha2 = "0.11"
 
 [features]
-micro-git = ["dep:shipper-git"]
 micro-encrypt = []
 micro-cargo = ["dep:shipper-cargo"]
 micro-plan = ["dep:shipper-plan", "micro-types"]
@@ -69,7 +68,7 @@ micro-webhook = []
 micro-types = ["micro-webhook", "micro-encrypt", "micro-config"]
 micro-config = ["micro-types"]
 micro-parallel = ["micro-types", "micro-plan", "dep:shipper-engine-parallel"]
-micro-all = ["micro-git", "micro-encrypt", "micro-cargo", "micro-plan", "micro-webhook", "micro-types", "micro-config", "micro-parallel"]
+micro-all = ["micro-encrypt", "micro-cargo", "micro-plan", "micro-webhook", "micro-types", "micro-config", "micro-parallel"]
 
 [lints]
 workspace = true
